pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "cems-nuclei"
version = "2.0.1"
description = "Python wrapper around NUCLEI's functionality."
dependencies = ["aiohttp>=3.11.16,<4", "pyjwt>=2.6.0,<3"]
requires-python = ">=3.11"
license = { file = "LICENSE.txt" }
readme = "README.md"
keywords = ["API"]

[project.urls]
repository = "https://github.com/cemsbv/nuclei"

[project.optional-dependencies]
client = ["orjson>=3,<4"]
test = ["coveralls", "pytest"]
docs = [
  "Sphinx==8.2.3",
  "sphinx-autodoc-typehints==3.1.0",
  "sphinx_rtd_theme",
  "asteroid-sphinx-theme==0.0.3",
  "anywidget==0.9.18",
  "numpy>=2.2.4,<3",
  "ipython>=9.1.0,<10",
]
lint = ["black", "isort"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
files = ["nuclei"]
mypy_path = 'src'
namespace_packages = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_subclassing_any = true
no_implicit_optional = true
disallow_untyped_defs = true
strict_concatenate = true
disallow_any_generics = false
disallow_untyped_calls = true
warn_return_any = false
no_implicit_reexport = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
  "numpy.*",
  "pytest.*",
  "setuptools.*",
  "jwt.*",
  "IPython.*",
  "orjson.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests/*"]
disallow_untyped_defs = false
